Batch `update_last`
