+++ docker-compose.yml
@@ -84,8 +84,7 @@ services:
     volumes:
       - ./database/__init__.py:/app/database/__init__.py:ro
       - ./database/asyncdatabase/__init__.py:/app/database/asyncdatabase/__init__.py:ro
-      - ./service/chat/__init__.py:/app/service/chat/__init__.py:ro
-      - ./service/chat/auth.py:/app/service/chat/auth.py:ro
+      - ./service/chat:/app/service/chat:ro
       - ./service/chat/container/mongooseim.toml:/mongooseim.template.toml:ro
       - ./service/chat/container/init.sql:/init.sql:ro
       - ./service/chat/container/init-db.sh:/init-db.sh:ro
