+++ service/chat/username/__init__.py
@@ -1,0 +1,3 @@
+class Username:
+    def __init__(self):
+        self.username = None
